backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
google-generativeai==0.8.3
scikit-learn==1.6.1
xgboost==2.1.3
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
